Updated Tests for HL APIs
