+++ tests/cloud/high_level_testing.json
@@ -1,6 +1,6 @@
 {
     "license": "SEE LICENSE IN LICENSE FILE",
-    "model_name" : ["gpt2","lu-vae/llama-68m-fft"],
+    "model_name" : ["gpt2"],
     "num_cores" : [16],
     "prompt" : ["My name is"],
     "prompts_txt_file_path" : ["examples/prompts.txt"],
